Stop config task generating templates
We want templates to be committed to the source repo so that local
customization is easier and to allow for better discovery of
configuration file templates, i.e. no magical generation at
configuration time.

Removes references to local overrides in the README but leaves the code
in place for backwards compatibility.