Recursively process directories with .dice files
This removes the previously deprecated functionality but adds full
recursive directory descent with scenario coverage. Only files with the
'.dice' extension will be processed.

This is a minimally invasive set of changes. More refactoring is
probably possible but I've left this asside to get the new functionality
up and running as quickly as possible.